+++ src/Components/Categories/CategoriesNew.css
@@ -1,0 +1,17 @@
+.categories-new {
+  width: 40%;
+}
+
+.categories-new__button-group {
+  margin-top: 98px;
+  display: flex;
+  justify-content: flex-end;
+}
+
+.categories-new__button-group button:last-child {
+  margin-left: 30px;
+}
+
+.categories-new .text-box {
+  margin-bottom: 45px;
+}
